Refactor diff markdown rendering modules
